Arreglando tipo de dato double

+++ app/src/main/java/com/ucne/adielgarcia_p2_p2/data/remote/dto/DepositoDto.kt
@@ -1,7 +1,6 @@
 package com.ucne.adielgarcia_p2_p2.data.remote.dto
 
 import com.ucne.adielgarcia_p2_p2.utils.DateSerializer
-import kotlinx.serialization.Contextual
 import kotlinx.serialization.Serializable
 import java.util.Date
 
@@ -12,5 +11,5 @@ data class DepositoDto (
     val fecha: Date,
     val idCuenta: Int,
     val concepto: String,
-    val monto: Int
+    val monto: Double
 )
