+++ app/src/main/java/com/ucne/adielgarcia_p2_p2/presentation/depositos/form/FormViewModel.kt
@@ -92,7 +92,7 @@ class FormViewModel @Inject constructor(
              fecha = _uiState.value.fecha.formatAsDate() ?: Date(),
              idCuenta = _uiState.value.idCuenta.toInt(),
              concepto = _uiState.value.concepto,
-             monto = _uiState.value.monto.toInt()
+             monto = _uiState.value.monto.toDouble()
          )
 
         navigateBack()
